Objects now reset to their original positions when moving from test mode back to object placement mode. Player is repositioned everytime a test run is started for a room, stopping the player from going out of bounds when not rendering. Water now slows the player down. Speed is now set back to normal when leaving water tiles to normal tiles

+++ NoCodeGameEditor/Player.h
@@ -14,7 +14,7 @@ public:
 	Player();
 	~Player();
 
-	int m_defaultSpeed;
+	float m_defaultSpeed;
 	bool m_weaponEnabled = false;
 	sf::Sprite m_playerSprite;
 	sf::Texture m_playerTexture;
